requirements.txt
discord
py-cord
pymongo
pymongo[srv]
dis-uniter
requests
pillow